chapter_002/src/test/java/ru/job4j/chapter_006/io/ConfigTest.java
package ru.job4j.chapter_006.io;


import org.junit.Test;

import static org.hamcrest.Matchers.is;
import static org.junit.Assert.assertThat;

public class ConfigTest {

    @Test
    public void whenPairWithoutComment() {
        String path = "./data/pair_without_comment.properties";
        Config config = new Config(path);
        config.load();
        assertThat(
                config.value("name"),
                is("Petr Arsentev")
        );
    }
    @Test
    public void whenPairWithComment() {
        Config config = new Config("./data/app.properties");
        config.load();
        assertThat(
                config.value("hibernate.connection.url"),
                is("jdbc:postgresql://127.0.0.1:5432")
        );
    }
}
